feat: Add lightweight video thumbnail for UI
Introduce a thumbnail field to `FileItem` to store a base64 encoded first frame of videos. This allows the UI to render an `<img>` tag instead of a `<video>` tag for videos, significantly improving performance and reducing memory usage when displaying lists of video files. The `generateMetadataForFile` service has been updated to extract and return this thumbnail.